Learned features: refactor, split screen, TODO
Here I learned some features of the PyCharm again, What I learned here is the PEP 8 style of coding which is most known style for writing The CODE. I also learned the features like How to watch the split screen, How to refactor multiple function, understood the structure. With this I completed the basic journey of the PyCharm.